requirements.txt
flask
Flask-Compress
gunicorn
APScheduler==3.10.4
requests
beautifulsoup4
lxml
python-dateutil
pytz
python-dotenv
pyyaml
